requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-surprise>=1.1.3
scikit-learn>=1.3.0
scipy>=1.10.0
pytest>=7.4.0
pytest-cov>=4.1.0

# PyTorch with CUDA support
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
# For CPU-only: pip install torch torchvision torchaudio
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# PyTorch Geometric (install after PyTorch)
torch-geometric>=2.3.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6
pyyaml>=6.0
httpx>=0.27.0

# UI
streamlit>=1.31.0
requests>=2.31.0
